refactor: oss配置相关重构调整
