--- spring-boot-extension-autoconfigure/src/main/java/com/livk/autoconfigure/oss/client/OSSClientFactory.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.livk.autoconfigure.oss.client;
-
-/**
- * The interface Oss client factory.
- *
- * @param <T> the type parameter
- * @author livk
- */
-public interface OSSClientFactory<T> {
-
-	/**
-	 * Instance t.
-	 * @param endpoint the endpoint
-	 * @param accessKey the access key
-	 * @param secretKey the secret key
-	 * @param region the region
-	 * @return the t
-	 */
-	T instance(String endpoint, String accessKey, String secretKey, String region);
-
-	/**
-	 * Prefix string.
-	 * @return the string
-	 */
-	String name();
-
-}
