+++ spring-boot-extension-autoconfigure/src/main/java/com/livk/autoconfigure/oss/support/AbstractService.java
@@ -17,7 +17,7 @@
 package com.livk.autoconfigure.oss.support;
 
 import com.livk.autoconfigure.oss.OSSProperties;
-import com.livk.autoconfigure.oss.client.OSSClientFactory;
+import com.livk.autoconfigure.oss.factory.OSSClientFactory;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
@@ -25,6 +25,8 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.support.GenericApplicationContext;
 import org.springframework.core.ResolvableType;
+
+import java.util.Optional;
 
 /**
  * The type Abstract service.
@@ -41,17 +43,25 @@ public abstract non-sealed class AbstractService<T> implements OSSOperations, Ap
 	@Override
 	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
 		OSSProperties properties = applicationContext.getBean(OSSProperties.class);
-		OSSClientFactoryPatternResolver resolver = new OSSClientFactoryPatternResolver();
-		OSSClientFactory<T> factory = resolver.loader(properties.getPrefix());
-		this.client = factory.instance(properties.getEndpoint(), properties.getAccessKey(), properties.getSecretKey(),
-				properties.getRegion());
 
-		if (applicationContext instanceof GenericApplicationContext context) {
-			ResolvableType resolvableType = ResolvableType.forInstance(client);
-			BeanDefinitionBuilder definitionBuilder = BeanDefinitionBuilder.rootBeanDefinition(resolvableType,
-					() -> client);
-			AbstractBeanDefinition beanDefinition = definitionBuilder.getBeanDefinition();
-			context.registerBeanDefinition(factory.name(), beanDefinition);
+		ResolvableType type = ResolvableType.forClass(OSSClientFactory.class);
+		Optional<OSSClientFactory<T>> optional = applicationContext.<OSSClientFactory<T>>getBeanProvider(type)
+			.orderedStream()
+			.filter(factory -> factory.name().equals(properties.getType()))
+			.findFirst();
+
+		if (optional.isPresent()) {
+			OSSClientFactory<T> factory = optional.get();
+			this.client = factory.instance(properties.getEndpoint(), properties.getAccessKey(),
+					properties.getSecretKey(), properties.getRegion());
+
+			if (applicationContext instanceof GenericApplicationContext context) {
+				ResolvableType resolvableType = ResolvableType.forInstance(client);
+				BeanDefinitionBuilder definitionBuilder = BeanDefinitionBuilder.rootBeanDefinition(resolvableType,
+						() -> client);
+				AbstractBeanDefinition beanDefinition = definitionBuilder.getBeanDefinition();
+				context.registerBeanDefinition(factory.name(), beanDefinition);
+			}
 		}
 	}
