+++ spring-boot-extension-tests/oss-spring-boot-test/minio-spring-boot-test/src/main/resources/application.yml
@@ -2,4 +2,5 @@ spring:
   oss:
     access-key: admin
     secret-key: 1375632510
-    url: minio:http://livk.com:9000
+    endpoint: http://livk.com:9000
+    type: minio
