+++ spring-boot-extension-tests/oss-spring-boot-test/minio-spring-boot-test/src/test/java/com/livk/oss/minio/controller/OssControllerTest.java
@@ -57,7 +57,8 @@ class OssControllerTest {
 	static void properties(DynamicPropertyRegistry registry) {
 		registry.add("spring.oss.access-key", minio::getUserName);
 		registry.add("spring.oss.secret-key", minio::getPassword);
-		registry.add("spring.oss.url", () -> "minio:" + minio.getS3URL());
+		registry.add("spring.oss.endpoint", minio::getS3URL);
+		registry.add("spring.oss.type", () -> "minio");
 	}
 
 	@Autowired
